// lib/types.ts
export interface ProjectMeta {
  title: string
  description: string
  date: string
  slug: string
  draft?: boolean
}

export interface WritingMeta {
  title: string
  description: string
  date: string
  slug: string
  draft?: boolean
}

export interface ResearchMeta {
  title: string
  description: string
  date: string
  slug: string
  draft?: boolean
}